backend/requirements_rag.txt
# RAG AI Server Dependencies (Port 8000)
# Install: pip install -r requirements_rag.txt

# FastAPI
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.7
python-dotenv==1.0.1
pydantic==2.6.1

# LangChain
langchain==0.1.20
langchain-community==0.0.38
langchain-google-genai>=4.2.1

# Google Gemini
google-generativeai>=0.8.6

# PDF processing
pypdf==4.2.0

# Local embeddings (no API key needed)
sentence-transformers>=2.7.0

# Vector store
faiss-cpu==1.8.0
